+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 
 # fqt-todo changelog
 
+## v1.7
+
+optimization search query and query count
+fix small screen display problem
+
+2019-10-30
+
 ## v1.6
 
 config data make more specification